double-pendulum-robot/requirements.txt
numpy>=1.26
scipy>=1.12
sympy>=1.12
matplotlib>=3.8
pyyaml>=6.0
pytest>=8.0
